Document rule 5107
Summary:
Rule 5107 is identical to 5101, it just covers data explicitly annotated on API Framework types. It was originally separated out to make it easier to productionize 5101. At this point, we have the ability to productionize at least a subset, so i'm adding documentation for it. But there's not point in duplicating the content in 5101, so the docs just point users there.

Note: This could be done with a client side redirect, but I wanted people to be explicitly aware that 5101 and 5107 are the same, so they don't get confused when they land on the 5101 page.

Reviewed By: dkgi

Differential Revision: D29239363

fbshipit-source-id: 9dd4fe59bd7778e7f7472e05f6e66446c8e7dbe3