Add a model decorator for skipping inlining.
Summary:
Problem: Inlining a decorator can increase the trace length by up to 2 for flows through the original function. There's no good solution for this since the alternative increases trace length by 1 but leads to unfriendly UX like `foo` pointing to the decorator body in the Zoncolan UI.

Our tradeoff is to allow the former kind of inlining but to opt out decorators that are "safe". That way, we don't have to modify our heuristic of using a short trace length as a proxy for usefulness since we will ignore most decorators.

Going for opt-out instead of opt-in so that we can check new decorators added in a diff while still ignoring existing decorators.

Remaining work:
+ Actually use the above model decorator to skip inlining.

Reviewed By: arthaud

Differential Revision: D29215192

fbshipit-source-id: c1f0c0e07fc480e698feac2431a5324f26b43c38