Make analyzeCommand parse models before calling static analysis.
Summary:
Problem: We want to know which decorators to skip inlining before we start the static analysis. So, we need to parse the models from the model sources and learn which decorators need to be skipped.

This diff just does the parsing up front in `analyzeCommand` (only when inlining is enabled). Will use the actual info later in the stack.

Reviewed By: arthaud

Differential Revision: D29215189

fbshipit-source-id: 145136fff063974bf0e468833b3f14576637f148