Move sanitizers out of modes
Summary:
Sanitizers are currently encoded in the `mode` part of a model.
Problem: some combinations of sanitizers with modes are not supported, such as having both sanitizers and `SkipDecoratorWhenInlining`.
Solution: move sanitizers out of modes, and make the mode a set of modes.
This diff does the first part, i.e moving sanitizers out of the mode.

Reviewed By: dkgi

Differential Revision: D29245704

fbshipit-source-id: f6663b4fe2236e7e627b1457fc5143840ff0aac1