Infer: replace `-use-v2` with `-infer-mode` [3/3] cleanup
Summary:
## About this commit

I'm working on a stack to replace the `-use-v2` flag with
`-infer-mode interprocedural`, see below for the rationale.

This is a breaking change, which means it can't be released in
one shot, we need to do it in 3 parts:
 - make a forward/backward compatible ocaml cli
 - flip the python client
 - remove the dead ocaml flag

This commit is the last, it cleans up the no-longer-used
`-use-v2` flag.

## Why make this change?

This stack is an admission that interprocedural infer remains
somewhat on the horizon at this point:
- pysa jobs already routinely time out, so potentially having
  slower analysis (due to pre-running a full typecheck, building
  a callgraph, and doing two iterations even with our dummy call models)
  is risky
- we still have quite a bit of work to do to make a true call model,
  which means we get no benefits from interprocedural performance hits
  in the meantime

But we still ought to roll out the new interprocedural logic in a
timely manner:
- we definitely would like to get rid of AnalysisError.t as the
  mechanism for propagating type inference
- the new python client is cleaner than the old one, and we have
  bug reports on the python side code genration piling up

So, with this commit I want to move toward being able to use
typeInfrenceData on non-interprocedural infer, which would allow us

Reviewed By: dkgi

Differential Revision: D29216014

fbshipit-source-id: fe6bcb5650baf885109ebbef7047bc7bc7919609