Stop dumping all toplevel string value assignments in debug
Summary:
I'm trying to use debug mode to look into memory usage to debug another problem, and found that for some reason when running `--debug`, my type check process hung indefinitely in the middle of parsing sources.

This was bizarre because when `--debug` was not passed in, the process finished without a problem; parsing sources only took about 6 seconds and env building another 6 after.

It turns out that our error logging in preprocessing can be quite unreasonable -- since toplevel assignments /could/ be type aliases, we try to parse the value of every toplevel assignment and qualify it. When this fails (in most cases), we print the entire value as a "bad annotation" that we couldn't parse.

Not only does this make it really hard to actually use this logging to identify string annotations that are invalid, but this results in printing out entire files sometimes where multi-line strings are used on the toplevel. An example file: P423985833 (line 814 through the end of file is printed as a single 'bad annotation')

I didn't get to the bottom of exactly which file was causing the indefinite (30 min+, at least) hanging behavior, but removing the logging of failed to parse string annotations fixed the problem.

To avoid losing signal on string annotations that we are 100% sure do need to be a valid type annotation, we need to distinguish between strings to definitely qualify (ie., explicit type aliases, return annotation strings, etc.), attempt-to-qualify-but-failures-expected (toplevel assignments), and do not qualify.

Reviewed By: grievejia

Differential Revision: D29247498

fbshipit-source-id: 984e861393ec3e9f923bb933b246e68864ae67c0