Refactor type checking in the interprocedural analysis [1/n]
Summary:
Problem:
Writing a Pysa model for a function removes its obscureness, which means it won't be taint-in-taint-out.
This is confusing, especially for people simply trying to add a source or sink on a function.

To fix that, we need the model parsing to keep the obscureness of models, which means we need the set of stubs during the first step of model parsing. This is hard to accomplish now with the recent changes introduces model initialization in two steps.
For pysa, the first step is called directly in `analyzeCommand.ml` before we event built the set of stubs. The second step is called in `staticAnalysis.ml`.

To ease the transition, this stack aims at refactoring all steps of the interprocedural analysis into small functions that can be re-used.

Reviewed By: pradeep90

Differential Revision: D29320880

fbshipit-source-id: aca6e2fcffd26f9cef40a2b2ba8e8197fdf26706