Infer: local infer-v2 [2] Create infer_for_module
Summary:
This stack is building out a non-interprocedural v2 infer
that we can use to push the v2 python client (which is decoupled
from AnalysisError) out without having to worry about the perf
issues that interprocedural infer might cause.

This commit creates a local infer for one module, returning a
list of `TypeInferenceData.LocalResult` values for all the defines
in a module. The logic is lifted almost exactly from the existing
`run` function in `Inference.ml` (which we will be able to remove
entirely once this is released).

Reviewed By: shannonzhu

Differential Revision: D29337511

fbshipit-source-id: 426eb3e0c2431ce9fa90a15b28d4f5778670fa07