Infer: local infer-v2 [3] extract environment building logic
Summary:
This stack is working on a v2 of local infer, that uses the
`TypeInferenceData` types that we've created for interprocedural
infer but doesn't require an interprocedural fixpoint.

This commit preps for making a v2 infer in `Service.Infer` by
extracting environment-building logic, which is a big chunk of the
total code, into a function that will be sharable across the
legacy `AnalysisError`-based infer and the new infer.

Reviewed By: shannonzhu

Differential Revision: D29337544

fbshipit-source-id: d0ab81ed3edc85c5f82045a7b7598ee5702505a7