Infer: local infer-v2 [5] Add local support to python client
Summary:
Add an `--interprocedural` flag to `infer-v2`. The default value is
`False`, in which case we call the non-interprocedural v2 ocaml code.

The only difference from a python client's point of view is a flag
sent to ocaml; the output format is identical (and at the moment the
actual results are identical except that the interprocedural code
doesn't yet filter out external sources).

Reviewed By: shannonzhu

Differential Revision: D29343872

fbshipit-source-id: 291023773c8ef73bd6e1eab3ee3baea033a01e7e