Infer: local infer-v2 [1] Set up wrapper around inference.ml
Summary:
In this stack of commits, I'm going to implement a local-only infer-v2 that
spits out results using TypeInferenceData.GlobalResult like the interprocedural
infer. This will allow us to shift to the v2 python client and start rolling out
bugfixes without having to decide whether interprocedural infer is worth the
wall time costs (it's quite a bit slower).

In this diff, I kick off the process by extracting a new `infer_for_define`
into `typeInference/typeInferenceLocal.ml` that we can share between the
local-only and interprocedural codebases; this wraps the legacy
`interprocedural.ml` `infer_for_define` with the logic to produce a
`TypeInferenceData.LocalResult` from the list of `AnalysisError`.

Reviewed By: shannonzhu

Differential Revision: D29337494

fbshipit-source-id: 1e1872bbb386cea8f9063fad828c8c16cdafd14f