Infer: local infer-v2 [4] Add `-infer-mode local` support
Summary:
This stack adds a non-interprocedural infer that supports the v2
client by using `TypeInference.Data.GlobalResult` rather than a list
of `AnalysisError.t` to report results. We want to ship this rapidly
so that we can roll out the v2 python client and start squashing
code-generation bugs ASAP.

This commit is the final commit for the ocaml side of this: it adds
an `infer_v2` function to `Service.Infer`, and a new function
`run_infer_local` to `inferCommand.ml` which we use whenever the
`-infer-mode local` flag is passed.

Reviewed By: shannonzhu

Differential Revision: D29337549

fbshipit-source-id: 7bacdeef4bd227fc5eeacb12cc9ba99ef47e796c