Fetch callables before initializing models [6/n]
Summary:
See the diff at the bottom of the stack for more context.
We need access to the stubs when parsing models in order to preserve the obscureness of models.
To do that, we need to fetch callables before initializing models, and then pass functions and stubs to the initialization function.
The next diff will finally use stubs to preserve the obscureness of models.

Reviewed By: dkgi

Differential Revision: D29349522

fbshipit-source-id: 7a12dc33292c6ba4f7dac377b0d6c01de268f1de