Infer: move all shared utilities to infer_v2.py
Summary:
At this point the ocaml client is ready for us to start moving
from infer to infer-v2 as the "main" infer implementation.

It therefore makes sense to move all the code shared between
the implementations into infer-v2 so that the code and tests
won't be affected when we're ready to remove the legacy infer
logic.

There's no logic change here, only
- moving the code
- updating annotations (use list/set/| instead of List/Set/Union)

Reviewed By: shannonzhu

Differential Revision: D29370084

fbshipit-source-id: 66100d5678ec0afecc8be867de5f88458df31e24