Join of invariant containers
Summary:
Our typeorder join was order-dependent when passing in parametrics containing Any.

In other words, join of `List[int]` and `List[Any]` was `List[Any]`, but a join of `List[Any]` and `List[int]` was `List[int]`.

In practice, this is because `Any` is less than *and* greater than every other type, so our call to `always_less_or_equal` would default to a simple less or equal comparison when it finds no free type variables, and Any < int is true - so is int < Any.

This is pretty inconsistent with the behavior of join when parametrics aren't involved. We have match cases that ensure that a join of `int` and `Any`, in either order, always returns `Any`.

Making sure parametrics do the same will also help solve the root cause of the pyre OOM when we start inferring mutable container literals with more precision. More detail on that coming in a future diff.

Reviewed By: pradeep90

Differential Revision: D29305748

fbshipit-source-id: 49c8853e833df978599a4be27c10f61320a0722e