Infer: refactor AnnotateModuleInPlace
Summary:
I want to start building a test suite for the merge-stub-with-code
logic (which is mostly coming from libcst, but to debug issues
with generated infer code we need to understand it).

This commit modifies the code to support this by isolating file
loading (which I don't want in unit tests) from the core logic
that operates on strings. I add a simple placeholder unit test.

Reviewed By: shannonzhu

Differential Revision: D29397517

fbshipit-source-id: 5c4b1d1089bfb0d624cbd004076d9510d646b5fc