Uninitialized local: suppress walrus-operator false positives
Summary:
The fix-point computation of initialized variables is currently statement level. This is a quick fix to suppress reads of things defined using a walrus operator and read in the same expression.

It comes at the cost of also suppressing errors in an expression like `x = f(x)` and  `y := x and x := 3`. That said, we are general prioritizing reducing false positives even if they come at the cost of a few false negatives -- so till we have a proper fix it makes sense to do this.

Reviewed By: pradeep90

Differential Revision: D29403178

fbshipit-source-id: f698a72932c5fb53d124392ac5acbfab79a7482a