Add an annotation `@SkipObscure` to remove the obscureness of a callable
Summary:
The previous diff preserves the obscureness of callables being modelled in `.pysa` files.
This leads to a lot of functions being marked as obscure by default, such as `list.append`, `dict.get`, etc.

This diff adds a taint annotation `SkipObscure` that can be used in a `.pysa` file to prevent a callable from being obscure, and hence taint-in-taint-out for all parameters.

A follow-up diff will update our taint stubs to reflect this.

Reviewed By: dkgi

Differential Revision: D29371018

fbshipit-source-id: fa2bba9cb511891aa80a5ba06511e99cec30deff