Keep the obscureness of modelled callables
Summary:
When writing models on obscure functions, we implicitly remove the obscureness of it.
This leads to confusion for users that do not expect the function to be obscure in the first place.
This patch keeps the obscureness of modelled callables.

Note that this will result in a lot of `via:obscure` breadcrumbs being present on traces, since a lot of standard library functions with models will become obscure, such as `int.__add__` or `list.append`. In a follow-up, we will add a decorator to remove the obscureness explicitly, which should address that problem.

Reviewed By: dkgi

Differential Revision: D29364718

fbshipit-source-id: 52ef0673e83e0c05faca94d9bda9169f6078bf5a