Add `@SkipObscure` on our taint stubs
Summary:
We are now preserving the obscureness of callables by default. See the previous diffs for more context.
They are still cases where we want to explicitly remove the obscure-ness so that a function is not treated as taint-in-taint-out.
This diff adds the `SkipObscure` to all callables with `TaintInTaintOut` annotation that are defined in type stubs.

Reviewed By: dkgi

Differential Revision: D29372079

fbshipit-source-id: 2631cde4922eac2102fe8fd94063154c65cbcdc1