Uninitialized local: support for assert False
Summary: CFG building takes care of raising AssertionError, but does not treat an "assert False" the same way. This is the source of discrepancy between type check and uninitialized local check. Type check appears to have some custom logic for "assert False", whereas uninitialized local check relies on the CFG for handling this.

Reviewed By: dkgi

Differential Revision: D29438278

fbshipit-source-id: ae42e9d2123885cd592dec6385ffbfc2e53f7437