Preserve prefix arguments passed to original function.
Summary:
Problem: Right now, we replace `original(request, *args, **kwargs)` with all parameters from the original signature: `original(request, x, y)`.

However, that is a problem when a decorator passes on a local variable `x`. When we start supporting `wraps`, qualification will end up treating the local variable `x` differently from the parameter `x` and this will lead to unintuitive flows.

So, preserve the prefix arguments as they are and just replace `*args, **kwargs` with whatever we inferred as their precise values.

Reviewed By: arthaud

Differential Revision: D29433431

fbshipit-source-id: 1457f26017d7160807a0bae1178b727c4725f62e