Ignore decorator if not all calls are identical.
Summary:
Problem: We want to ignore decorators that call the original function multiple times with different arguments.

This is not really a pattern we see in existing decorators, but is a useful invariant in other changes we want to make down the line. For example, when inlining a decorator, if we want to infer the contents of args and kwargs up fornt, we don't want to raise an error there about being unable to do so because two calls have different values for args and kwargs.

Reviewed By: arthaud

Differential Revision: D29433433

fbshipit-source-id: 02a3b40e5b222c943a4236ada27ac5d1bab56a9e