Infer: fix sanitized-vs-unsanitized bug
Summary:
The better error message introduced in the previous diff
demonstrated a bug in the new infer code: it turns out that
we are inconsistently creating MissingParameterAnnotation errors
sometimes with sanitized names and other times with unsanitized.

The easiest solution is to push sanitization down into the
parameters handling of typeInferenceData - sanitizing an already
sanitized name is a no-op so this can't cause problems. So we
- create the parameters map using sanitized names rather than
  unsanitized as the keys
- sanitize names before updating with inferred annotations

Reviewed By: pradeep90

Differential Revision: D29569542

fbshipit-source-id: 47efb1c785932a38edab953501c914f25da3357d