Infer: migrate parameter annotation unit tests for functions
Summary:
I'm separating functions and methods into separate tests and
diffs to keep things smaller and easier to navigate; the old
test suite combined both into a single huge test.

I already migrated the first couple of tests in the previous diff
to verify a bugfix, but this commit migrates the remainder of the
function parameter inference tests.

Reviewed By: grievejia

Differential Revision: D29574315

fbshipit-source-id: 58051093b491ffb2e4731d551750685f42902ed0