Support new `Parameters(<taint>, where=...)` clause
Summary:
**Motivation**
We want to be able to specify conditions on what parameters to taint, such as parameters matching a specific type annotation, or parameters matching/not matching a particular name. One current use case is to exclude parameters of type `IGWSGIRequest` when migrating the existing `get_POST_annotated_sources` Python model generator.

**Syntax**
```
ModelQuery(
    name = "get_POST_annotated_sources",
    find = "functions",
    where = [Decorator(name.matches("api_view"), arguments.contains("POST"))],
    model = [
        Parameters(
            TaintSink[Test],
            where=[
                Not(AnyOf(
                    name.matches("self"),
                    name.matches("cls"),
                    type_annotation.matches("IGWSGIRequest"),
                    type_annotation.matches("HttpRequest"),
                ))
            ]
        )
    ]
)
```

And this also works:
```
ModelQuery(
    find = "methods",
    where = [parent.matches("C")],
    model = [
        Parameters(
            TaintSink[Test],
            where=[
                type_annotation.is_annotated_type()
            ]
        )
    ]
)
```
which will e.g. match argument `y` in `foo` here:
```
class C:
   def foo(self, x: str, y: Annotated[str, "foo"]): ...
```

**Description**
To achieve this in a more general way than just adding another `excludes_types` argument, this diff adds a `where` clause to the general `Parameters` model clause. The clauses supported in this `where` context are `AnyOf`, `Not`, `name` and `type_annotation`.

**Note**: This new `Parameters` with a `where` clause should also cover the existing `AllParameters`, etc. The goal is to eventually remove all of those and only have this one `Parameters` model clause.

Reviewed By: pradeep90

Differential Revision: D29551843

fbshipit-source-id: 270625769e721bdc4fc1662bb5676ee6290cf8b1