Tidy up logic for determining find clause kind
Summary:
Refactored the functions that determine find clause kind to not use strings and use proper types instead.

Also add a check to allow `parent` clause only when find clause is "methods" or "attributes". We were previously missing this check for the `parent` clauses that the find clause kind was either `"attributes"` or `"methods"`. Let's add this in, and add a unit test for it.

Also made a helper function for checking whether the find clause kind is a callable type (i.e. methods or functions). This will be helpful for setting up the `"parameter"` find clause kind in future diffs on this stack.

Reviewed By: pradeep90

Differential Revision: D29550566

fbshipit-source-id: eac0e6d454fa13259315eb2eb61bfbd070706589