Make `where` clause optional in `Parameters`
Summary:
This makes the `where` clause option. When no `where` clause is specified `Parameters` should just behave like the existing `AllParameters`.

Syntax:
```
ModelQuery(
  find = "functions",
  where = [
    name.matches("test8_")
  ],
  model = [
    Parameters(TaintSource[Test])
  ]
)
```

With this change, we should be able to cover all existing `PositionalParameter`, `NamedParameter`, and `AllParameters` use cases with the new `Parameter(where=...)` clause. We should be able to migrate existing models to the new syntax and remove the existing clauses.

Reviewed By: pradeep90

Differential Revision: D29591901

fbshipit-source-id: 92123c98906c2f1230eb57575113ffb9617c24f7