Implement PEP 647 -- User-Defined Type Guards
Summary:
== PEP specification
https://www.python.org/dev/peps/pep-0647/#specification

== Implementation
We have verified that we cover the various cases mentioned in the specification. In particular, we have verified that:
* TypeGuard accepts exactly one type argument
* Return statements within a type guard function should return bool values, and we are verifying that all return paths return a bool
* Type is narrowed for first parameter of functions
* Type is narrowed for the second parameter of methods
* User-defined type guards that take more than one argument are supported (note that no type narrowing is applied to those additional argument expressions)
* User-defined type guards that are generic functions are supported (both with one argument, and more than one)
* Do not require that the return type be strictly narrower
* Do not narrow type of the negative branch (`else`)
* Works with overloads and callbacks as expected
* Narrowing of dictionary to TypedDict using a user-defined TypeGuard works

== Background technical information
In Pyre, narrowing of `if condition` is handled via Assert `condition`. This is where some of the existing type-narrowing logic exists, and where we have added support for user-defined TypeGuard "type-narrowing" (in quotes, because it isn't always a strict narrowing -- PEP requires to adhere to whatever the user specifies).

== Follow-up

The spec if underspecified in certain scenarios, and in the follow-up we will be documenting/discussing how Pyre should handle these as well as other edge cases we can think of.

Reviewed By: pradeep90

Differential Revision: D29595687

fbshipit-source-id: 5d007e1e437ba681b5281baa1d447506614fc2c2