Infer: add lines between classes in stubs
Summary:
As I was working to debug a case where legacy infer adds annotations
that infer-v2 does not, I compared the generated stubs and noticed
that infer-v2 stubs lacked an extra newline between classes which
make them hard to read.

This commit adds the space, and a test verifying the change.

Reviewed By: pradeep90

Differential Revision: D29654206

fbshipit-source-id: 72676695d6e1ea830bc0eed7e6511ba0f911e830