Infer: add typeInferenceLocal.mli
Summary:
I've been working without .mli files in some of the unstable
infer modules, because (particularly on big stacks) they can
really slow down work when developing brand new code where
reviewers are likely to request a bunch of changes to the
interface.

But at this point the interface seems stable enough that it
shouldn't be a big problem.

NOTE: cool trick for this is
```
opam install ocaml-print-intf
dune exec -- ocaml-print-intf path/to/file.ml
```

Reviewed By: grievejia

Differential Revision: D29650653

fbshipit-source-id: 4aa1298e182cd96b349a8e1b9bb79604c9912ca1