Infer bugfix: skip duplicate defines
Summary:
There exist libraries that use decorators and involve multiple
definitions of the same function. The only standard one I can think
of is `typing.overloads`, but it's possible to write other tools that
do the same thing, for example I encountered a library that uses
`compile` decorators which I think are being used for a kind of
pattern matching.

The overloads case probably wouldn't matter for infer since an
overload has an empty body, but that is not true of all other cases.
As a result, we need to handle this situation.

The most correct answer seems to be that we should skip inference
entirely whenever we see a duplicate function name, because any library
doing this has semantics that infer won't understand (with the possible
exception of overloads, but in that case the user needs to be manually
specifying types anyway).

Reviewed By: pradeep90

Differential Revision: D29653455

fbshipit-source-id: 14aa93b2de5ba4dc7197e04d66511fe8ef431a1d