Infer bugfix: patch path handling
Summary:
After stepping through the path handling in painful detail, I
think I understand what infer does to translate buck paths into
actual code paths. This diff fixes `infer-v2` to behave the same way.

I added a docstring to the new `transform_raw_path` method explaining
what it does and why.

I also forced the `type_directory` to always be absolute, because
that was hiding incorrect logic in the tests. It's easier reason
about the code if the type_directory is *always* absolute than if
it can be relative.

Reviewed By: pradeep90

Differential Revision: D29654203

fbshipit-source-id: d265c7db297c0ead3dbe66c4fab510cd66e0852b