Infer: be more intelligent with NoneType
Summary:
Prior to this commit it was an open TODO to improve None handling
I figured I should just patch it before the next release and attempt
to deploy. The reason we want to skip inferring NoneType is that
a field which is always None probably wouldn't exist in the first place;
there's almost certainly some logic we can't see that sets the field
(currently infer very rarely does inference for nonlocals).

The approach here is to filter NoneType inferences for globals
and attributes at the GlobalResults stage, not the LocalResults
stage, because we *don't* want to suppress local inference in
cases where the type may get widened - at the moment this happens
when a variable is initialized both in a class toplevel and a
constructor.

**Context / why now?**

The `infer-v2` command is now working well enough
that I was able to compare it head-to-head against
`infer` for a couple serious projects, and the only
differences I could find were:
- `infer-v2` skipped adding annotations to some duplicate
  defines, which is expected; we didn't have the ability
  to do it correctly before so I removed it altogether
  for now, and it's pretty rare anyway.
- `infer-v2` was adding some annotations of `None` to
  attributes and globals that `infer` was skipping. I actually
  can't figure out where the difference came from, it was an open
  TODO item to filter these in the backend but maybe legacy infer
  is filtering in python (I hunted but couldn't find where)

Reviewed By: grievejia

Differential Revision: D29671492

fbshipit-source-id: 09f31725b79365c549057badcc7554d21f2961e7