Infer + sqlalchemy [0]: Add option to dequalify annotations
Summary:
In order to roll out the new infer completely, I need to
fix the fact that `support_sqlalchemy.py`, one of the subcommands
of `pyre-upgrade`, is rewriting inference errors to fix path
qualification issues.

That kind of operation is very much interfering with private
parts of the infer interface (the ocaml to python bridge should
definitely not be considered public), so the solution seems to
be to push the dequalify logic down into `infer` as an option
via a flag.

Hence this diff:
- update the old `dequalify_and_fix_pathlike` to more clearly
  state what transformation it applies to annotations, and
  rename it to `sanitize_annotations`
- add a `dequalify_all` option to `sanitize_annotations` that
  performs the transformation from `support_sqlalchemy.py`.
  Lift the unit tests from there, and organize things a bit more.
- add a `dequalify` flag to the commandline that will trigger
  us to sanitize with the `dequalify_all` option turned on

Reviewed By: dkgi

Differential Revision: D29651678

fbshipit-source-id: 4fb42ee7d6456af2fd4b3f11e67fbaef56c9929a