Infer: remove dupicate validation in __init__
Summary:
I moved the validation of annotate_from_existing_stubs vs
in_place to the top of __init__, that's my preferred
style for readability.

But there were two issues: I forgot to modify the check at
top to work with boolean in_place (it used to be an optional
list), and (hiding that problem) I had forgotten to remove
the check at the bottom of `__init__`.

Reviewed By: dkgi

Differential Revision: D29672509

fbshipit-source-id: 635696f23d927d213f2936806d97dc77ec1a5f19