Infer rollout [0] use infer_v2 code for `pyre infer`
Summary:
In this commit, I copy/paste the implementation of `pyre infer-v2`
for `pyre infer`. As of this commit `infer-v2` is still needed because
we need to migrate all use cases before landing, then flip them back
before removing the `infer-v2` command altogether.

I'll follow up with a big cleanup diff removing the legacy infer
codebase, but I'd rather keep the actual cutover diff small; the
only cleanup in this diff is of the integration tests (I
removed the legacy test entirely, and renamed the v2 test to just
be `InferTest`).

I did discover a place where I was accidentally using legacy Infer
in the infer_v2 unit tests, so I had to do a tiny patch of that code.

Reviewed By: kbansal

Differential Revision: D29638883

fbshipit-source-id: 90e00d5fd523f88336db7ff0fd9b6909ea8d0d26