Infer rollout [3]: remove `inference` from LegacyError
Summary:
We no longer need to support the `inference` field in
LegacyError - with infer gone it is no longer used, and later
in this stack I'm going to remove it from the ocaml side.

NOTE: We will need to land this diff and update the client
prior to pushing (or at least releasing) logic that removes the ocaml
side because it's fine to have extra fields python ignores, but
python would crash if it tried to parse errors with no
`inference_information` entry.

Reviewed By: pradeep90

Differential Revision: D29651283

fbshipit-source-id: b7353a395a4e83f72b680749757148c35e00f94f