Added `BroadcastError` error for reporting `BroadcastError` types
Summary:
+ Problem: We want to eliminate the `BroadcastError` type in favor of a `BroadcastError` error, which is raised upon detecting the aforementioned type.
+ Background: Currently, we have a `BroadcastError` type which is returned by any `Broadcast` errors that we have. We don't want this type to stick around though, since it's kind of not nice (and will cause a lot of further errors to be reported, potentially), and we want it to be more of a marker than a true "type". This is a change to directly improve the user experience of `Broadcast` in error situations.
+ Solution: Added a new error, `BroadcastError`. This is raised whenever a `forward_callable` result type potentially contains a `BroadcastError` type. We silently change the type of the offending expression to be `Any`, as well, so that `BroadcastError` doesn't stick around.

**Why add `BroadcastError` as a first class type?**
`BroadcastError` needs to be propagated all the way from `local_replace`. One approach is to change all the affiliated functions from `local_replace` to `forward_callable` to return a sum type of a `Type.t` versus an `Error` - this is a rather bulky change. What we can do is instead use the existing types to propagate a `Type.t` up to `forward_callable`, and then erase all traces of it and report the error.

**Why not use Bottom instead of BroadcastError?**
`Bottom` does not behave in a very user-friendly way - it is possible to keep using the expression (even with a return type of `Bottom`), and Pyre will silently allow it, instead of reporting anything. It's better to not use `Bottom`, to allow the possibility of better user experience.

Reviewed By: pradeep90

Differential Revision: D29669897

fbshipit-source-id: 9eddeec0a0e28b337d97c9a124a9bec2ba4cd1f8