Represent explicit modules as SourcePaths in error data
Summary:
Make the full source path object available for error objects that point to a module. This will allow us to have access to more information about the module, such as whether it is a stub, and if so where to go look for it.

 ---

While looking into this, I discovered that the lookup for `module_exists` wasn't quite the same as `lookup_search_path`. `module_exists` also checks an `implicit_submodule` hash table, see: D17075336 (https://github.com/facebook/pyre-check/commit/ad9569ecae5a5af40bfcda64a59473f4587ce0be)

This is because some modules are explicit, i.e, there is a path to a concrete file, and these can have a source path representation.

Other modules do exist, but they don't have a concrete file, and are "implicitly" defined because a submodule of them is explicit. If we have `a.b.c.py`, then `a`, `a.b` are implicit and `a.b.c` is explicit.

Some other things I considered for this diff:
1. Passing in a resolution to error message creation, which we can use to look up the source path of qualifiers we care about on the fly. Decided against this because semantically, we shouldn't need to be doing environment-based lookups to determine how to convert an error object into a string. I don't think we'll need this for future changes to error message quality, either.

2. Replacing Reference.t with SourcePath.t (turns out not possible to do correctly given implicit modules don't have a source path but can still be valid part of error). Then, considered using an `is_stub` boolean field instead of an `Explicit/Implicit` type variant. Decided against this because it hides away why we have to do two checks, `module_exists` and then `lookup_search_path` which already returns an optional. Would rather just encode in the data structure that the stub field is not relevant to all modules.

3. Trying to tackle other errors besides UndefinedX. By far the most common stub confusion I've seen involves missing definitions of attributes or globals in stub files. I think it'll add too much complexity to try to add an origin field to all other kinds of errors, ie., IncompatibleParameterType technically could be caused by an expected parameter type set in a stub. Eventually when we have better error trace support, where to go look for the expected type can be made clearer across the board.

Reviewed By: kbansal

Differential Revision: D29724942

fbshipit-source-id: aaed24a2660e77bababe392ef8f3c46a9a173f2b