Support stub messaging in class attribute error messages
Summary:
When we're erroring on an undefined attribute of a class, also pass in the source path of the module defining the class (if it's available - not handling edge cases like nested classes for now).

We can use this information to point people to the right stub file to look in if they find they're expecting an attribute to exist that Pyre doesn't know about, and the relevant class is defined in a .pyi.

Reviewed By: grievejia

Differential Revision: D29745271

fbshipit-source-id: ab25e0d85c14bab62cbf4c6099444d6de1f82861