Log when consolidation of existing configurations is not possible
Summary:
simple nfc.

Just wanted to add a line of logging because I thought a call to consolidate nested would work in a directory with multiple configs, and it exited silently.
Turns out I misread the paths and they're not nested after all. Would have appreciated a log saying so next time.

Reviewed By: grievejia

Differential Revision: D29809767

fbshipit-source-id: 3d49d8cad70d7f1e99ea1a44350f4aa88b5cdb58