Work around `re2` build issue on MacOS
Summary:
While working on pulling `pyre-ast` in as a dependency, I noticed that our opamcake build is failing for package `re2`: https://www.internalfb.com/intern/sandcastle/log/?instance_id=457949769&step_id=4706069902&step_index=3&name=Build

Apparently this is caused by a [known bug](https://github.com/janestreet/re2/issues/31) in OCaml version < 4.12 on MacOS. As the linked page mentioned, it is unlikely that the patch will be backported to prior version of OCaml. The best workaround we have atm is to either upgrade to 4.12 (which may lead to a perf regression that needs to be investigated first), or remove the extra `version` file ourselves manually. This diff takes the latter approach which is easier to do atm.

Reviewed By: pradeep90

Differential Revision: D29838284

fbshipit-source-id: c49a077accf31a930f974dd0c7a9775762709247