Pick up targets to migrate if any typing-related settings are present
Summary:
We care about removing type check settings that are being deprecated even if they are included in targets that don't also have a `check_types=True` line. Often this means these fields are no-ops and should be cleaned up.

To avoid touching more than we need to in a migration codemod, we can still only add targets to pyre configs if they include an explicit `check_types=True` (consistent with previous behavior).

Reviewed By: pradeep90

Differential Revision: D29833908

fbshipit-source-id: 7dc1b33f1c66fc6464f426015e2559d6719179b0