Change integration test to check an allowlist of keys
Summary: To prevent issues like T96156049 in the future, we need to ignore keys that are not salient to determining whether our our results have changed or not. To do that, we introduce an allowlist of the "code", "define", "description", and "path" keys, and only use those when comparing `results.json` files. I believe those keys will be sufficient to prove that we haven't stopped finding/started finding any issues relative to the golden master.

Reviewed By: stroxler

Differential Revision: D29831873

fbshipit-source-id: c12ee70eb44f87bf0ee4cc81e5172ed39ddbf195